requirements.txt
boto3
beautifulsoup4
requests
python-dateutil
newspaper3k
lxml_html_clean